+++ NEWS.md
@@ -1,7 +1,7 @@
 StMoMo v0.4.1.9000
 ----------------------------------------------------------------
 * Adjust random walk code to work with missing values correctly
-
+* Add new utility functions defining the plat model
 
 StMoMo v0.4.1
 ----------------------------------------------------------------